requirements.txt
# Python Backend Dependencies
# ML/Analytics Pipeline Architecture

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Scientific Computing & ML
numpy==1.26.2
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.3.2

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
